# apis/backend-api/partner/sdkwork-partner-backend-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Partner Backend API",
    "version": "0.1.0",
    "description": "Backend administration API for the multi-level partner (agent) management system: levels and commission ratios, partners, join fees, customer bindings, commission events and settlements, ledger, withdrawals, and stats.",
    "x-sdkwork-api-authority": "sdkwork-partner-backend-api",
    "x-sdkwork-sdk-family": "sdkwork-partner-backend-sdk",
    "x-sdkwork-owner": "sdkwork-partner"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
...
  "paths": {
...
    "/backend/v3/api/partners/commission_config": {
      "get": {
        "operationId": "commissionConfig.retrieve",
        "summary": "Retrieve the commission configuration",
        "tags": [
          "partners"
        ],
        "x-sdkwork-owner": "sdkwork-partner",
        "x-sdkwork-api-authority": "sdkwork-partner-backend-api",
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-permission": "commerce.partner.read",
        "responses": {
          "200": {
            "$ref": "#/components/responses/CommissionConfigItemEnvelopeResponse"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      },
      "patch": {
        "operationId": "commissionConfig.update",
        "summary": "Update the commission configuration",
        "tags": [
          "partners"
        ],
        "x-sdkwork-owner": "sdkwork-partner",
        "x-sdkwork-api-authority": "sdkwork-partner-backend-api",
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-permission": "commerce.partner.manage",
        "responses": {
          "200": {
            "$ref": "#/components/responses/CommissionConfigItemEnvelopeResponse"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCommissionConfigUpdateRequest"
              }
            }
          }
        }
      }
    },
...
    "/backend/v3/api/partners/{partnerId}": {
      "get": {
        "operationId": "partner.retrieve",
        "summary": "Retrieve a partner",
        "tags": [
          "partners"
        ],
        "x-sdkwork-owner": "sdkwork-partner",
        "x-sdkwork-api-authority": "sdkwork-partner-backend-api",
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-permission": "commerce.partner.read",
        "responses": {
          "200": {
            "$ref": "#/components/responses/PartnerItemEnvelopeResponse"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "name": "partnerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "int64",
              "pattern": "^-?[0-9]+$",
              "x-sdkwork-int64-string": true
            },
            "description": "Partner id."
          }
        ]
      },
      "patch": {
        "operationId": "partner.update",
        "summary": "Update a partner",
        "tags": [
          "partners"
        ],
        "x-sdkwork-owner": "sdkwork-partner",
        "x-sdkwork-api-authority": "sdkwork-partner-backend-api",
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-permission": "commerce.partner.manage",
        "responses": {
          "200": {
            "$ref": "#/components/responses/PartnerItemEnvelopeResponse"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPartnerUpdateRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "partnerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "int64",
              "pattern": "^-?[0-9]+$",
              "x-sdkwork-int64-string": true
            },
            "description": "Partner id."
          }
        ]
      }
    },
...
    "/backend/v3/api/partners/{partnerId}/stats": {
      "get": {
        "operationId": "stats.retrieve",
        "summary": "Retrieve a partner's aggregated stats",
        "tags": [
          "partners"
        ],
        "x-sdkwork-owner": "sdkwork-partner",
        "x-sdkwork-api-authority": "sdkwork-partner-backend-api",
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-permission": "commerce.partner.read",
        "responses": {
          "200": {
            "$ref": "#/components/responses/PartnerStatItemEnvelopeResponse"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "name": "partnerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "int64",
              "pattern": "^-?[0-9]+$",
              "x-sdkwork-int64-string": true
            },
            "description": "Partner id."
          }
        ]
      }
    },
...
    "/backend/v3/api/partners/applications/{applicationId}": {
      "get": {
        "tags": [
          "partners"
        ],
        "x-sdkwork-owner": "sdkwork-partner",
        "x-sdkwork-api-authority": "sdkwork-partner-backend-api",
        "x-sdkwork-api-surface": "backend-api",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "operationId": "applications.retrieve",
        "summary": "Retrieve a partner join application",
        "x-sdkwork-permission": "commerce.partner.read",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "int64",
              "pattern": "^-?[0-9]+$",
              "x-sdkwork-int64-string": true
            },
            "description": "Join application id."
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AdminPartnerApplicationItemEnvelopeResponse"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "SdkWorkApiResponse": {
        "type": "object",
        "required": [
          "code",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "const": 0
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "CommissionConfigItemEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "item"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/CommissionConfigItem"
                  }
                }
              }
            }
          }
        ]
      },
      "AdminCommissionConfigUpdateRequest": {
        "type": "object",
        "required": [
          "enabled",
          "usageSettlementEnabled",
          "rechargeEnabled",
          "maxCommissionDepth",
          "currency",
          "minWithdrawalAmount"
        ],
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Global commission settlement toggle."
          },
          "usageSettlementEnabled": {
            "type": "boolean",
            "description": "Commission on usage settlement revenue."
          },
          "rechargeEnabled": {
            "type": "boolean",
            "description": "Commission on recharge revenue."
          },
          "maxCommissionDepth": {
            "type": "string",
            "description": "Max ancestor depth (0 = unlimited).",
            "format": "int64",
            "x-sdkwork-int64-string": true,
            "pattern": "^-?[0-9]+$"
          },
          "currency": {
            "type": "string",
            "description": "Commission currency code."
          },
          "minWithdrawalAmount": {
            "type": "string",
            "description": "Minimum withdrawal amount."
          },
          "profitMarginRatio": {
            "type": "string",
            "description": "Platform gross profit margin (percent, e.g. 40.00). Customer revenue commissions are profit-based: the allocation base equals revenue × margin; join-fee commissions use the full join fee."
          }
        }
      },
...
      "PartnerItemEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "item"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/PartnerItem"
                  }
                }
              }
            }
          }
        ]
      },
...
      "AdminPartnerUpdateRequest": {
        "type": "object",
        "required": [
          "name",
          "levelNo",
          "status"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Partner display name."
          },
          "contactName": {
            "type": "string",
            "description": "Contact name."
          },
          "phone": {
            "type": "string",
            "description": "Contact phone."
          },
          "email": {
            "type": "string",
            "description": "Contact email."
          },
          "levelNo": {
            "type": "integer",
            "description": "Partner level number.",
            "format": "int32"
          },
          "parentPartnerId": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Parent partner id (null = top level, clears the current parent).",
            "format": "int64",
            "x-sdkwork-int64-string": true
          },
          "userAccountId": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Bound IAM user account id (null = unbind).",
            "format": "int64",
            "x-sdkwork-int64-string": true
          },
          "status": {
            "type": "string",
            "description": "Partner status.",
            "enum": [
              "PENDING",
              "ACTIVE",
              "SUSPENDED",
              "CLOSED"
            ]
          },
          "remark": {
            "type": "string",
            "description": "Remark."
          }
        }
      },
...
      "PartnerStatItemEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "item"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/PartnerStatItem"
                  }
                }
              }
            }
          }
        ]
      },
...
      "AdminPartnerApplicationItemEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "item"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/AdminPartnerApplicationItem"
                  }
                }
              }
            }
          }
        ]
      },
...
    },
    "responses": {
...
      "CommissionConfigItemEnvelopeResponse": {
        "description": "Success",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CommissionConfigItemEnvelope"
            }
          }
        }
      },
...
      "PartnerItemEnvelopeResponse": {
        "description": "Success",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PartnerItemEnvelope"
            }
          }
        }
      },
...
      "PartnerStatItemEnvelopeResponse": {
        "description": "Success",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PartnerStatItemEnvelope"
            }
          }
        }
      },
...
      "AdminPartnerApplicationItemEnvelopeResponse": {
        "description": "Success",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AdminPartnerApplicationItemEnvelope"
            }
          }
        }
      },
...
    }
  },
...
}